+++ CHANGELOG.md
@@ -10,6 +10,10 @@ The format is based on [keep a changelog](http://keepachangelog.com/) and this p
 - Fix NakamaSocket.add_matchmaker_party_async() and the tests for it
 - Fix MatchData.op_code type in schema to TYPE_INT
 
+### Added
+
+- Add support for receiving binary data in "NakamaRTAPI.MatchState"
+
 ## [3.1.0] - 2022-04-28
 
 ### Added